+++ src/main/java/com/example/springrider/domain/order/service/OwnerOrderService.java
@@ -1,7 +1,7 @@
 package com.example.springrider.domain.order.service;
 
 import com.example.springrider.aop.StoreOwnerCheck;
-import com.example.springrider.config.PasswordEncoder;
+import com.example.springrider.global.security.PasswordEncoder;
 import com.example.springrider.domain.common.exception.ExceptionCode;
 import com.example.springrider.domain.common.exception.InvalidRequestException;
 import com.example.springrider.domain.order.dto.request.CancelOrderRequestDto;
